chore: remove default index page from public middleware
